Refactor Map Helpers for readability
