+++ app/src/UI/Map2/helpers/constants.ts
@@ -1,0 +1,1 @@
+export const FALLBACK_COLOR = 'red';
